refactor: AuTab0Viewerのコンポーネント化と共有ロジックの抽出
- AuTab0Viewer.tsxを複数の専門コンポーネントに分割
- 共通のナビゲーション・ハイライトロジックをカスタムフック`useAuTab0Navigation`に抽出して共有化
- `CompactAccordion`と`ViewerOptionRow`を使用して、表示密度を向上（縦方向を詰める）
- AGENT.mdの規約に従い、Propsインターフェースの命名や制御構文の記述を修正
- 状態を持つコンポーネント（feature）と持たないコンポーネント（components）を適切に分離
- ユニットテストおよびE2Eテストにて正常動作を確認済み
